+++ lib/models/playlist.dart
@@ -1,5 +1,4 @@
 import 'package:youtube_data_api/models/thumbnail.dart';
-
 class PlayList {
   ///Youtube playlist id
   String? playListId;
@@ -25,9 +24,9 @@ class PlayList {
 
   factory PlayList.fromMap(Map<String, dynamic>? map) {
     List<Thumbnail> thumbnails = [];
-    map?['playlistRenderer']['thumbnails'][0]['thumbnails']
+    map?['playlistRenderer']['thumbnails']
         .forEach((thumbnail) {
-      thumbnails.add(Thumbnail(url: thumbnail['url']));
+      thumbnails.add(Thumbnail(url: thumbnail['thumbnails'][0]['url'], width: thumbnail['thumbnails'][0]['width'], height: thumbnail['thumbnails'][0]['height']));
     });
     return PlayList(
         playListId: map?['playlistRenderer']['playlistId'],