+++ lib/models/video.dart
@@ -1,3 +1,4 @@
+import 'package:youtube_data_api/models/thumbnail.dart';
 class Video {
   ///Youtube video id
   String? videoId;
@@ -14,37 +15,50 @@ class Video {
   ///Youtube video views
   String? views;
 
-  String? thumbnail;
+  ///Youtube video thumbnail
+  List<Thumbnail>? thumbnails;
 
   Video(
-      {this.videoId, this.duration, this.title, this.channelName, this.views, this.thumbnail});
+      {this.videoId, this.duration, this.title, this.channelName, this.views, this.thumbnails});
 
   factory Video.fromMap(Map<String, dynamic>? map) {
-    if(map?.containsKey("videoRenderer")?? false){
-      ///Trending and search videos
+    List<Thumbnail>? thumbnails;
+    if(map?.containsKey("videoRenderer") ?? false){
+      //Trending and search videos
       var lengthText = map?['videoRenderer']?['lengthText'];
       var simpleText =
       map?['videoRenderer']?['shortViewCountText']?['simpleText'];
+      thumbnails = [];
+      map?['videoRenderer']
+      ['thumbnail']['thumbnails']
+          .forEach((thumbnail) {
+        thumbnails!.add(Thumbnail(url: thumbnail['url'], width: thumbnail['width'], height: thumbnail['height']));
+      });
       return Video(
           videoId: map?['videoRenderer']?['videoId'],
           duration: (lengthText == null) ? "Live" : lengthText?['simpleText'],
           title: map?['videoRenderer']?['title']?['runs']?[0]?['text'],
           channelName: map?['videoRenderer']['longBylineText']['runs'][0]['text'],
+          thumbnails: thumbnails,
           views: (lengthText == null)
               ? "Views " +
               map!['videoRenderer']['viewCountText']['runs'][0]['text']
               : simpleText);
     } else if (map?.containsKey("compactVideoRenderer")?? false){
-      ///Related videos
+      //Related videos
+      thumbnails = [];
+      map?['compactVideoRenderer']
+      ['thumbnail']['thumbnails']
+          .forEach((thumbnail) {
+        thumbnails!.add(Thumbnail(url: thumbnail['url'], width: thumbnail['width'], height: thumbnail['height']));
+      });
       return Video(
         videoId: map?['compactVideoRenderer']['videoId'],
         title: map?['compactVideoRenderer']
         ?['title']?['simpleText'],
         duration: map?['compactVideoRenderer']
         ?['lengthText']?['simpleText'],
-        thumbnail: map?['compactVideoRenderer']
-        ['thumbnail']['thumbnails'][1]
-        ['url'],
+        thumbnails: thumbnails,
         channelName: map?['compactVideoRenderer']
         ?['shortBylineText']?['runs']
         ?[0]?['text'],
@@ -54,12 +68,19 @@ class Video {
     } else if(map?.containsKey("gridVideoRenderer")?? false) {
       String? simpleText = map?['gridVideoRenderer']
       ['shortViewCountText']?['simpleText'];
+      thumbnails = [];
+      map?['gridVideoRenderer']
+      ['thumbnail']['thumbnails']
+          .forEach((thumbnail) {
+        thumbnails!.add(Thumbnail(url: thumbnail['url'], width: thumbnail['width'], height: thumbnail['height']));
+      });
       return Video(
           videoId: map?['gridVideoRenderer']['videoId'],
           title: map?['gridVideoRenderer']['title']['runs'][0]
           ['text'],
           duration: map?['gridVideoRenderer']['thumbnailOverlays'][0]
           ['thumbnailOverlayTimeStatusRenderer']['text']['simpleText'],
+          thumbnails: thumbnails,
           views: (simpleText != null) ? simpleText : "???"
       );
     }